Initial dev branch update

+++ server/api/override/delete.post.ts
@@ -2,13 +2,20 @@ import { Override } from "../../../db/models";
 
 export default defineEventHandler(async (event) => {
   const body = await readBody(event);
+  const checkResult = await userCheck(event, body.SystemId, undefined, undefined);
+  if (checkResult.BoundaryRoleId && checkResult.BoundaryRoleId !== 4) {
+    const override = await Override.destroy({
+      where: {
+        SystemId: body.SystemId,
+        StigDatumId: body.StigDatumId,
+      },
+    });
 
-  const override = await Override.destroy({
-    where: {
-      SystemId: body.SystemId,
-      StigDatumId: body.StigDatumId,
-    },
-  });
-
-  return override;
+    return override;
+  } else {
+    throw createError({
+      statusCode: 401,
+      statusMessage: "Insufficient Permissions.",
+    });
+  }
 });
